Tests/LibWeb/Text/expected/css/style-invalidation/user-action-pseudo-class-targeted-invalidation.txt

plain text hover: styleInvalidations=0, elementStyleRecomputations=0, elementStyleNoopRecomputations=0
link hover: styleInvalidations=1, elementStyleRecomputations=2, elementStyleNoopRecomputations=1
link color: rgb(0, 128, 0)
active state: fullStyleInvalidations=0, elementStyleRecomputations=8
